user_profile: extract profile buttons to their own component
as that's already a good chunk on its own
